Move client-ui-lib into shared-text
