--- examples/data-objects/client-ui-lib/.eslintrc.js
@@ -1,13 +1,0 @@
-/*!
- * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
- * Licensed under the MIT License.
- */
-
-module.exports = {
-    "extends": [
-        "@fluidframework/eslint-config-fluid"
-    ],
-    "parserOptions": {
-        "project": ["./tsconfig.json", "./src/test/tsconfig.json"]
-    },
-}
