--- examples/data-objects/client-ui-lib/.npmignore
@@ -1,6 +1,0 @@
-nyc
-*.log
-**/*.tsbuildinfo
-src/test
-dist/test
-**/_api-extractor-temp/**
