--- examples/data-objects/client-ui-lib/.vscode/settings.json
@@ -1,3 +1,0 @@
-{
-    "typescript.tsdk": "node_modules/typescript/lib"
-}
