--- examples/data-objects/client-ui-lib/README.md
@@ -1,1 +1,0 @@
-# @fluid-example/client-ui-lib
