--- examples/data-objects/client-ui-lib/package.json
@@ -1,105 +1,0 @@
-{
-  "name": "@fluid-example/client-ui-lib",
-  "version": "0.59.3000",
-  "private": true,
-  "description": "Fluid client UI",
-  "homepage": "https://fluidframework.com",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/microsoft/FluidFramework.git",
-    "directory": "examples/data-objects/client-ui-lib"
-  },
-  "license": "MIT",
-  "author": "Microsoft and contributors",
-  "sideEffects": false,
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "scripts": {
-    "build": "concurrently npm:build:compile npm:lint",
-    "build:compile": "npm run tsc && npm run build:test",
-    "build:full": "npm run build",
-    "build:full:compile": "npm run build:compile",
-    "build:test": "tsc --project ./src/test/tsconfig.json",
-    "clean": "rimraf dist *.tsbuildinfo *.build.log",
-    "eslint": "eslint --format stylish src",
-    "eslint:fix": "eslint --format stylish src --fix --fix-type problem,suggestion,layout",
-    "lint": "npm run eslint",
-    "lint:fix": "npm run eslint:fix",
-    "test": "npm run test:mocha",
-    "test:coverage": "nyc npm test -- --reporter xunit --reporter-option output=nyc/junit-report.xml",
-    "test:mocha": "mocha --recursive dist/test -r node_modules/@fluidframework/mocha-test-setup --unhandled-rejections=strict",
-    "test:mocha:verbose": "cross-env FLUID_TEST_VERBOSE=1 npm run test:mocha",
-    "tsc": "tsc",
-    "tsfmt": "tsfmt --verify",
-    "tsfmt:fix": "tsfmt --replace"
-  },
-  "nyc": {
-    "all": true,
-    "cache-dir": "nyc/.cache",
-    "exclude": [
-      "src/test/**/*.ts",
-      "dist/test/**/*.js"
-    ],
-    "exclude-after-remap": false,
-    "include": [
-      "src/**/*.ts",
-      "dist/**/*.js"
-    ],
-    "report-dir": "nyc/report",
-    "reporter": [
-      "cobertura",
-      "html",
-      "text"
-    ],
-    "temp-directory": "nyc/.nyc_output"
-  },
-  "dependencies": {
-    "@fluidframework/common-utils": "^0.32.1",
-    "@fluidframework/container-definitions": "^0.48.1000",
-    "@fluidframework/core-interfaces": "^0.43.1000",
-    "@fluidframework/datastore-definitions": "^0.59.3000",
-    "@fluidframework/ink": "^0.59.3000",
-    "@fluidframework/map": "^0.59.3000",
-    "@fluidframework/merge-tree": "^0.59.3000",
-    "@fluidframework/protocol-definitions": "^0.1028.1000",
-    "@fluidframework/request-handler": "^0.59.3000",
-    "@fluidframework/runtime-definitions": "^0.59.3000",
-    "@fluidframework/runtime-utils": "^0.59.3000",
-    "@fluidframework/sequence": "^0.59.3000",
-    "@fluidframework/undo-redo": "^0.59.3000",
-    "debug": "^4.1.1",
-    "react": "^16.10.2",
-    "react-dom": "^16.10.2"
-  },
-  "devDependencies": {
-    "@fluidframework/build-common": "^0.23.0",
-    "@fluidframework/eslint-config-fluid": "^0.28.2000-0",
-    "@fluidframework/mocha-test-setup": "^0.59.3000",
-    "@rushstack/eslint-config": "^2.5.1",
-    "@types/debug": "^4.1.5",
-    "@types/jsdom": "^12.0.0",
-    "@types/mocha": "^8.2.2",
-    "@types/node": "^14.18.0",
-    "@types/react": "^16.9.15",
-    "@typescript-eslint/eslint-plugin": "~5.9.0",
-    "@typescript-eslint/parser": "~5.9.0",
-    "concurrently": "^6.2.0",
-    "cross-env": "^7.0.2",
-    "eslint": "~8.6.0",
-    "eslint-plugin-editorconfig": "~3.2.0",
-    "eslint-plugin-eslint-comments": "~3.2.0",
-    "eslint-plugin-import": "~2.25.4",
-    "eslint-plugin-jest": "~26.1.3",
-    "eslint-plugin-mocha": "~10.0.3",
-    "eslint-plugin-promise": "~6.0.0",
-    "eslint-plugin-react": "~7.28.0",
-    "eslint-plugin-tsdoc": "~0.2.14",
-    "eslint-plugin-unicorn": "~40.0.0",
-    "jsdom": "^12.0.0",
-    "mocha": "^8.4.0",
-    "nyc": "^15.0.0",
-    "rimraf": "^2.6.2",
-    "typescript": "~4.1.3",
-    "typescript-formatter": "7.1.0"
-  }
-}
