--- examples/data-objects/client-ui-lib/src/test/tsconfig.json
@@ -1,20 +1,0 @@
-{
-    "extends": "@fluidframework/build-common/ts-common-config.json",
-    "compilerOptions": {
-        "rootDir": "./",
-        "outDir": "../../dist/test",
-        "types": [
-            "mocha"
-        ],
-        "declaration": false,
-        "declarationMap": false
-    },
-    "include": [
-        "./**/*"
-    ],
-    "references": [
-        {
-            "path": "../.."
-        }
-    ]
-}
