--- examples/data-objects/client-ui-lib/tsconfig.json
@@ -1,16 +1,0 @@
-{
-    "extends": "@fluidframework/build-common/ts-common-config.json",
-    "exclude": [
-        "src/test/**/*"
-    ],
-    "compilerOptions": {
-        "rootDir": "./src",
-        "outDir": "./dist",
-        "composite": true,
-        "types": ["react", "react-dom"],
-        "jsx": "react"
-    },
-    "include": [
-        "src/**/*"
-    ]
-}
