+++ examples/data-objects/shared-text/src/client-ui-lib/controls/cursor.ts
