+++ examples/data-objects/shared-text/src/client-ui-lib/controls/dockPanel.ts
