+++ examples/data-objects/shared-text/src/client-ui-lib/index.ts
