+++ examples/data-objects/shared-text/src/client-ui-lib/text/index.ts
