+++ examples/data-objects/shared-text/src/client-ui-lib/ui/geometry/index.ts
