+++ examples/data-objects/shared-text/src/client-ui-lib/ui/geometry/rectangle.ts
