+++ examples/data-objects/shared-text/src/client-ui-lib/ui/utils.ts
